users_to_scan table
In this update, I've changed both users fetcher and users_info fetcher.
I've added a new table to store users that they've been scanned and found.
This table has a field under the name last_seen. As user get discovered, the zero value assign to it, the user_info fetcher will use this table for his job. After fetching this user info, it changes the zero to one. Now we can make sure which users have been used and which ones not.